Verify the JWT when the WebSocket is opened
Moved verification of the signed JWT to when the socket is opened.
WebSocket support in the browser does not support adding additional
headers to the initial HTTP request, which means the JWT goes back
to the query string used to initiate the connection.

When the JWT is retrieved from the query string, it is verified
(potentially in parallel to receiving the ready message and
responding with the ack). If the JWT fails to verify, the connection
is closed with a reason (1008) and a message.

Signed-off-by: Erin Schnabel <schnabel@us.ibm.com>