add a test for mediator..
Adding a test to make it easier to add more tests.. this adds jmockit,
and junit to the gradle test step.

These tests do run ok in eclipse, although if using ibm jdk, you need to
add the -javaagent argument as it will instruct you when you try to run
them for the first time.

Remember to rebuild the eclipse project with gradle eclipse to have
eclipse pull in the new jars to use.

Signed-off-by: Ozzy Osborne <ozzy@ca.ibm.com>