Persist and move without kafka
Use the response from the player service to move the player between rooms (compensating
if there were conflicts). Send the service's UUID to player when requesting the
location update, so that location changes caused by this mediator can be ignored
(only relevant while the instance is running, otherwise meaningless data).

Signed-off-by: Erin Schnabel <schnabel@us.ibm.com>